Extract WikiFSLinks as new SPM target (module restructuring Phase 1)
Phase 1 of the module restructuring design (plans/module-restructure.md,
PR #534). Extracts the wiki-link grammar cluster from WikiFSCore into two
new SPM targets:

- WikiFSTypes (leaf, Foundation-only): PageID, ULID, ResourceKind,
  EmbedTarget, ParsedLink, MimeType — foundational types shared across
  all domain targets
- WikiFSLinks (depends on WikiFSTypes): 11 pure-logic link-grammar files
  moved from Sources/WikiFSCore/Links/

Resolves the circular WikiFSCore ↔ WikiFSLinks dependency by extracting a
WikiFSTypes leaf (the design doc's §5 cycle analysis identified only the
ParsedLink protocol-level cycle, but the implementation revealed a deeper
bidirectional coupling via PageID/ULID/ResourceKind/EmbedTarget/MimeType).
WikiFSCore re-exports both via @_exported import in ModuleExports.swift,
so existing importers need zero per-file import changes.

ParsedLink extracted from WikiLinkParser as a top-level public struct so
WikiStore.replaceLinks (WikiFSCore) can reference [ParsedLink] without
depending on WikiFSLinks. WikiLinkParser.ParsedLink → ParsedLink globally.

RelativeLinkRewriter.relativePath promoted to public (called cross-module
by SourceImageRewriter). Access-level tightening of ~22 internal-only types
deferred to a follow-up.

Build: 0 errors, 0 warnings. Tests: 2456 passed, 0 failures (fast tier).

Refs #532, #534.
